requirements.txt
flask
flask-cors
tensorflow
pillow
tqdm
torch
transformers
keras